requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.30.0

pydantic~=2.8.0

pandas~=2.2.0
numpy~=1.26.0

pyspark~=3.5.0

sentence-transformers~=3.0.0
faiss-cpu~=1.8.0

networkx~=3.2.0

pymongo~=4.8.0
redis~=5.0.0

httpx~=0.27.0

gradio~=4.44.0

pytest~=8.2.0
pydantic-settings~=2.5.0
